Reduce size of MADC channel type
Since MADC samples will be tagged with a
SourceMultiplexerOnReadoutSourceSelection coordinate, we can save a bit
of space by not assigning a 64 bit integer to each sample.

+++ include/halco/hicann-dls/vx/readout.h
@@ -24,8 +24,7 @@ struct GENPYBIND(inline_base("*")) ReadoutSourceSelectionOnDLS
 
 
 struct GENPYBIND(inline_base("*")) SourceMultiplexerOnReadoutSourceSelection
-    : public common::detail::
-          RantWrapper<SourceMultiplexerOnReadoutSourceSelection, uint_fast16_t, 1, 0>
+    : public common::detail::RantWrapper<SourceMultiplexerOnReadoutSourceSelection, uint8_t, 1, 0>
 {
 	constexpr explicit SourceMultiplexerOnReadoutSourceSelection(uintmax_t const val = 0)
 	    GENPYBIND(implicit_conversion) :
